services/message_queue/pyproject.toml
[project]
name = "message-queue"
version = "0.1.0"
description = "Turing Agents - Message Queue service"
readme = "README.md"
requires-python = ">=3.11,<3.14"
dependencies = [
  "aio-pika==9.4.1",
  "jsonschema==4.22.0",
  "pydantic==2.8.2",
  "python-dotenv==1.0.1",
  "opentelemetry-api==1.26.0",
  "opentelemetry-sdk==1.26.0",
  "opentelemetry-instrumentation==0.47b0",
  "prometheus-client==0.20.0",
  "httpx==0.27.2",
]

[project.optional-dependencies]
db = [
  "SQLAlchemy==2.0.35",
  "asyncpg==0.29.0",
  "sqlalchemy2-stubs==0.0.2a38",
]

[build-system]
requires = ["setuptools>=69", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
packages = ["libs", "scripts"]

[tool.pytest.ini_options]
addopts = "-q"
pythonpath = ["."]
testpaths = ["tests"]


